+++ tests/test_actor.py
@@ -20,3 +20,13 @@ async def test_status(actor):
 
     # command running + engineering mode + locked + 5 positioners + done
     assert len(actor.mock_replies) == 9
+
+
+async def test_info(actor):
+
+    command = await actor.invoke_mock_command('info')
+    assert command.status.did_succeed
+
+    data = actor.mock_replies[1:3]
+    assert 'version' in data[0]
+    assert 'config_file' in data[1]